main update

+++ _065_SOLID Principles/src/ShapesPrinter.java
@@ -1,11 +1,11 @@
 import java.util.List;
 
 public class ShapesPrinter {
-    public String json(int sum){
+    public String json(double sum){
         return "{shapesSum : %s}".formatted(sum);
   }
 
-    public String csv(int sum){
+    public String csv(double sum){
         return "shapes_Sum,%s}".formatted(sum);
     }
 }
